requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==2.3.0
Django==6.0.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
gspread==6.1.2
google-auth==2.29.0
